requirements.txt
distro==1.4.0
Django==2.2.5
django-cleanup==4.0.0
django-cors-headers==3.1.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.2
pandas==0.25.1
pandasql==0.7.3
Pillow==6.1.0
plotly==4.1.1
python-dateutil==2.8.0
pytz==2019.2
retrying==1.3.3
rpy2==2.9.5
six==1.12.0
SQLAlchemy==1.3.9
sqlparse==0.3.0
tabula-py==1.4.1
tzlocal==2.0.0
xlrd==1.2.0
